Rename BraveAppearanceHandler to BraveThemeHandler
To use BraveAppearanceHandler in many other places like welcome webui,
moved to parent folder and renamed to BraveThemeHandler.
