+++ browser/ui/BUILD.gn
@@ -70,14 +70,14 @@ source_set("ui") {
     "webui/brave_md_settings_ui.h",
     "webui/brave_new_tab_ui.cc",
     "webui/brave_new_tab_ui.h",
+    "webui/brave_theme_handler.cc",
+    "webui/brave_theme_handler.h",
     "webui/brave_web_ui_controller_factory.cc",
     "webui/brave_web_ui_controller_factory.h",
     "webui/brave_webui_source.cc",
     "webui/brave_webui_source.h",
     "webui/brave_welcome_ui.cc",
     "webui/brave_welcome_ui.h",
-    "webui/settings/brave_appearance_handler.cc",
-    "webui/settings/brave_appearance_handler.h",
     "webui/settings/brave_privacy_handler.cc",
     "webui/settings/brave_privacy_handler.h",
     "webui/settings/default_brave_shields_handler.cc",
