+++ browser/ui/webui/brave_md_settings_ui.cc
@@ -6,7 +6,7 @@
 
 #include "brave/browser/resources/grit/brave_settings_resources.h"
 #include "brave/browser/resources/grit/brave_settings_resources_map.h"
-#include "brave/browser/ui/webui/settings/brave_appearance_handler.h"
+#include "brave/browser/ui/webui/brave_theme_handler.h"
 #include "brave/browser/ui/webui/settings/brave_privacy_handler.h"
 #include "brave/browser/ui/webui/settings/default_brave_shields_handler.h"
 #include "chrome/browser/profiles/profile.h"
@@ -17,7 +17,7 @@ BraveMdSettingsUI::BraveMdSettingsUI(content::WebUI* web_ui,
                                      const std::string& host)
     : MdSettingsUI(web_ui) {
   web_ui->AddMessageHandler(std::make_unique<settings::MetricsReportingHandler>());
-  web_ui->AddMessageHandler(std::make_unique<BraveAppearanceHandler>());
+  web_ui->AddMessageHandler(std::make_unique<BraveThemeHandler>());
   web_ui->AddMessageHandler(std::make_unique<BravePrivacyHandler>());
   web_ui->AddMessageHandler(std::make_unique<DefaultBraveShieldsHandler>());
 }
